simplify animations
change animations so that instead of scale, translate, etc., being separately targetable by an animation, each object in the animations array must have completely specified transforms for the start and end of its animation. this simplifies things a lot because then we don't need to worry about deleting elements from the array dynamically and other bullshit. only one animation can run at a time. each loop we just look at the first item in the array and remove the front item when it's finished.