visualise the used-space grid
this allowed us to see and address a bug where labels being rendered off-screen were putting 1's in the usedSpace buffer where they shouldn't be. although we fixed that problem, the visualisation shows us that there is still a bug where labels rendered on one side of the screen wrap around and fill boxes in the used-space grid on the other side of the screen.