remove redundant copy
we were putting everything in one buffer before trying to send(). this meant that we wrote our header to a buffer and then copied the payload we got from the request handler to this buffer. now we avoid the copy by first sending the header and then sending the body. (the performance gains turned out to be negligible but it feels better this way.)