support anchors, unfortunately
maybe all regex is bad! but anchors especially so. it should be the default to match the whole string, because that's what's quick. but to get that behaviour you have to put in these ugly characters that have no relationship with what they mean: ^ and $. instead of that, we should tell people to put non-greedy match-all's at the start and end of the regex, i.e. `.*?`, if they don't want to match the whole. that might be ugly but it matches the ugliness of the execution that follows!