fix mainly two small sql errors
one was not expanding the booth_voronoi geometries to the envelope of the whole electorate. this meant that there were some gaps, particularly visible around bean.

the other was when we created the topo column in booths_22. we &&'d the topology face with the booth geometry, i.e. its point---instead of the point it should have been the voronoi as a whole. the result of this mistake was that we only held onto the part of the booth-voronoi multipolygon where the booth was actually located. this was particularly visible around cable beach in wa, where this portion of the multipolygon was a tiny fraction of the whole.