make Postgres_result faster
it is now an array of arrays with a single dict at the top level. it was previously an array of dicts. at the time i thought this was nice but profiling made it clear that there was a performance penalty to initialising all those dicts. in fact it was a pretty big penalty: the main voronoi query is now like 90% faster!!
i made a couple other changes that no doubt also helped:
- array_reserve-ing the cells when parsing the postgres result
- when parsing our own cached file, filling the postgres_result in with pointers into the file itself instead of making a copy for no reason