fix hash bug
whoops, hash_bytes was only returning a hash in the range of a u32 when hashing 4-byte values, because i made a mistake porting sean barrett's functions.

i also updated the seed to be a real 64-bit value, because we're not pretending to care about 32-bit functionality anymore. but that wasn't because of any bug and i'm not sure it changes a thing.