fix SIGPIPE bug
there was a bug where we would get a SIGPIPE and the whole program would crash if we tried to send() to an invalid socket. it turned out well before the send(), poll() was reporting these errors with the POLLHUP|POLLERR flags. so we just needed to check at these flags.