improve panel drag value
previously it was just boolean: the drag was either allowed or not allowed. if the drag would have brought the panel outside of the range of where it should be, it wasn't allowed. that didn't work because when you drag the panel up, often at the end of the drag you're asking it to go higher than it should be, and that means you want it to just go as high as it can go, but the effect of the boolean was that it would stop at the last frame where you happened to request a valid value.