fix browser display issues
chrome would sometimes display a column or row of white pixels to the right or bottom of the map, if window's size had a fractional component.
checking the difference this makes on firefox, i realised that it actually removes a small amount of blurriness which i'd sort of noticed but didn't even think could be fixed.