stop the http server from polling stdin
polling stdin stops the process from running in the background on linux (you can do it by redirecting /dev/null to stdin, but then it uses 100% of the cpu continually reading /dev/null).
also, poll() on windows doesn't support reading stdin anyway, so we've removed a barrier to making the http server portable.