strings: make copy_string() more flexible
i finally figured out how i want it to work. it's now consistent with get_string() and append_string() in that the string is a char_array. if you only want a char*, you can call it like this:

    char *copy = copy_string(string, length, context).data;

we also made a special copy_string_data() for map.c so that it doesn't have to depend on strings.h and array.h.

also i finally changed the behaviour of array_reserve() to just return early if we don't need to do anything, rather that asserting for dubious benefit.